Prepare feature file parsing in conversion command
